pyproject.toml
[tool.pdm]

[tool.pdm.scripts]
build = "python setup.py sdist bdist_wheel"
publish = "twine upload dist/* --skip-existing --verbose"

[tool.pdm.dev-dependencies]
dev = [
    "setuptools>=65.6.3",
    "wheel>=0.38.4",
]
[project]
name = "evue"
version = "0.1.3"
description = "Evue is the fastest way to build apps in Python or Javascript on windows/linux/macos/ios/andriod/rtos! Write once, run everywhere! "
authors = [
    {name = "dragondjf", email = "ding465398889@163.com"},
]
dependencies = [
    "flet>=0.2.4",
    "pyee>=9.0.4",
    "loguru>=0.6.0",
    "pyperclip>=1.8.2",
    "pillow>=9.3.0",
    "qrcode>=7.3.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
